breaks Feed component into smaller child components
