+++ client/src/components/Feed/PostList.jsx
@@ -1,0 +1,58 @@
+/*========== EXTERNAL MODULES ==========*/
+import React, { useState, useEffect } from 'react';
+import styled from 'styled-components';
+import Button from '@mui/material/Button';
+
+/*========== INTERNAL MODULES ==========*/
+import { Column } from '../../../public/stylesheets/styles.js';
+import Post from './Post.jsx';
+
+
+/*========== EXPORTS ==========*/
+export default function PostList({ handleClick, postings }) {
+  /*----- STATE HOOKS -----*/
+  // const [] = useState();
+
+
+  /*----- LIFESTYLE METHODS -----*/
+  // useEffect(() =>  {}, []);
+
+  /*----- EVENT HANDLERS -----*/
+
+
+
+
+  /*----- RENDER METHODS -----*/
+  const renderList = () => {
+    /*
+    NOTE:
+    - should render the basic information from the provided List prop,
+    - should conditionally render buttons specific to the parent page
+    - on Click -> sends the clicked posting information to the Details listing
+    */
+    if (postings) return postings.map(post => <Post handleClick={handleClick}/>)
+  }
+
+
+/*----- RENDERER -----*/
+return (
+  <ListSection>
+    {renderList()}
+  </ListSection>
+  )
+}
+
+
+
+
+/*========== STYLES ==========*/
+const ListSection = styled(Column)`
+  margin: 10px;
+  width: 48%;
+  height: 800px;
+  overflow: scroll;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+`;
